refactor: TransactorインターフェースをuseCase/portからdomain/repositoryへ移動
